Generate pdf code refactored
