+++ SprintManager.API/Program.cs
@@ -51,6 +51,8 @@ builder.Services.AddScoped<ITokenService, TokenService>();
 builder.Services.AddTransient<IEmailSender, EmailSender>();
 builder.Services.Configure<AuthMessageSenderOptions>(builder.Configuration.GetSection("SendGrid"));
 
+builder.Services.AddTransient<ISprintReportPdf, SprintReportPdf>();
+
 // Set inactivity timeout
 builder.Services.ConfigureApplicationCookie(o => {
     o.ExpireTimeSpan = TimeSpan.FromDays(5);
