+++ SprintManager.Application/Handlers/Sprints/GenerateSprintReportPdfHandler.cs
@@ -1,6 +1,5 @@
 ﻿using MediatR;
 using QuestPDF.Fluent;
-using QuestPDF.Helpers;
 using SprintManager.Application.DTOs;
 using SprintManager.Application.Interfaces;
 using SprintManager.Application.Queries.Sprints;
@@ -12,14 +11,17 @@ namespace SprintManager.Application.Handlers.Sprints
     {
         private readonly ISprintRepository _sprintRepository;
         private readonly IProjectRepository _projectRepository;
+        private readonly ISprintReportPdf _sprintReportPdf;
 
         public GenerateSprintReportPdfHandler(
             ISprintRepository sprintRepository,
-            IProjectRepository projectRepository
+            IProjectRepository projectRepository,
+            ISprintReportPdf sprintReportPdf
         )
         {
             _sprintRepository = sprintRepository;
             _projectRepository = projectRepository;
+            _sprintReportPdf = sprintReportPdf;
         }
 
         public async Task<PdfFileDTO> Handle(GenerateSprintReportPdfCommand request, CancellationToken cancellationToken)
@@ -30,38 +32,7 @@ namespace SprintManager.Application.Handlers.Sprints
 
             var project = await _projectRepository.GetByIdAsync(sprint.ProjectId);
 
-            var document = Document.Create(container =>
-            {
-                container.Page(page =>
-                {
-                    page.Size(PageSizes.A4);
-                    page.Margin(2, QuestPDF.Infrastructure.Unit.Centimetre);
-                    page.PageColor(Colors.White);
-                    page.DefaultTextStyle(x => x.FontSize(20));
-
-                    page.Header()
-                        .Text("Hello PDF!")
-                        .SemiBold().FontSize(36).FontColor(Colors.Blue.Medium);
-
-                    page.Content()
-                        .PaddingVertical(1, QuestPDF.Infrastructure.Unit.Centimetre)
-                        .Column(x =>
-                        {
-                            x.Spacing(20);
-
-                            x.Item().Text(Placeholders.LoremIpsum());
-                            x.Item().Image(Placeholders.Image(200, 100));
-                        });
-
-                    page.Footer()
-                        .AlignCenter()
-                        .Text(x =>
-                        {
-                            x.Span("Page ");
-                            x.CurrentPageNumber();
-                        });
-                });
-            });
+            var document = _sprintReportPdf.Compose();
 
             var fileName = $"{project?.Name.Replace(" ", "_")}_{sprint.SprintName.Replace(" ", "_")}_report".ToLower();
             var fileBytes = document.GeneratePdf();
