+++ SprintManager.Application/Interfaces/ISprintReportPdf.cs
@@ -1,0 +1,9 @@
+﻿using QuestPDF.Fluent;
+
+namespace SprintManager.Application.Interfaces
+{
+    public interface ISprintReportPdf
+    {
+        Document Compose();
+    }
+}
