+++ SprintManager.Infrastructure/Services/SprintReportPdf.cs
@@ -1,0 +1,52 @@
+﻿using QuestPDF.Fluent;
+using QuestPDF.Helpers;
+using SprintManager.Application.Interfaces;
+
+namespace SprintManager.Infrastructure.Services
+{
+    public class SprintReportPdf : ISprintReportPdf
+    {
+        public SprintReportPdf() 
+        { 
+            
+        }
+
+        public Document Compose()
+        {
+            var document = Document.Create(container =>
+            {
+                container.Page(page =>
+                {
+                    page.Size(PageSizes.A4);
+                    page.Margin(2, QuestPDF.Infrastructure.Unit.Centimetre);
+                    page.PageColor(Colors.White);
+                    page.DefaultTextStyle(x => x.FontSize(20));
+
+                    page.Header()
+                        .Text("Hello PDF!")
+                        .SemiBold().FontSize(36).FontColor(Colors.Blue.Medium);
+
+                    page.Content()
+                        .PaddingVertical(1, QuestPDF.Infrastructure.Unit.Centimetre)
+                        .Column(x =>
+                        {
+                            x.Spacing(20);
+
+                            x.Item().Text(Placeholders.LoremIpsum());
+                            x.Item().Image(Placeholders.Image(200, 100));
+                        });
+
+                    page.Footer()
+                        .AlignCenter()
+                        .Text(x =>
+                        {
+                            x.Span("Page ");
+                            x.CurrentPageNumber();
+                        });
+                });
+            });
+
+            return document;
+        }
+    }
+}
